requirements.txt
beautifulsoup4==4.4.1
colorama==0.3.5
Flask==0.10.1
Flask-Breadcrumbs==0.3.0
Flask-Login==0.3.2
Flask-Menu==0.5.0
Flask-PyMongo==0.4.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pymongo==3.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
termcolor==1.1.0
veryprettytable==0.8.1
Werkzeug==0.11.3
wheel==0.24.0
